update: separate render-worker.js from renderer.js
